requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==5.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.11
mariadb==1.1.14
mysqlclient==2.2.8
packaging==26.0
PyJWT==2.12.1
PyMySQL==1.1.2
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
urllib3==2.6.3
